Save and restore document window frame
Also add mechanism to save the UI state (scroll position, etc.).

In order to be able to retrieve the window frame in time to set it before the
whole document is read (reading the document is asynchronous), we save
the metadata in a new way.

Note: The previous file format was 100% CSV compatible. Now, you'll have to
remove the metadata before opening a file in a CSV editor in order to have
something consistent. Just remove everything up to the first non backslashed
semi-colon in the first line of the document.